Backend Send Link API Finished
API is actived by calling http://localhost:4000/subscriptionManagement/. We might want to include the user ID in it.
We need a page in frontend that calls that and render the result. Maybe into charts/list.
Another thing to add is to collect email sender & sent time when looking for link. we might want to decide if the link is valid based on the sent time. Link from emails sent half a year ago might be outdated.